# fix: prevent fast ascii comparison if char is not letter