--- a/pkg/logql/log/filter.go
+++ b/pkg/logql/log/filter.go
@@ -421,6 +421,9 @@ func (l equalFilter) String() string {
 }
 
 func newEqualFilter(match []byte, caseInsensitive bool) MatcherFilterer {
+	if caseInsensitive {
+		match = bytes.ToLower(match)
+	}
 	return equalFilter{match, caseInsensitive}
 }
 
@@ -441,7 +444,7 @@ func contains(line, substr []byte, caseInsensitive bool) bool {
 }
 
 // containsLower verifies if substr is a substring of line, with case insensitive comparison.
-// substr is expected to be in lowercase.
+// substr MUST be in lowercase before calling this function.
 func containsLower(line, substr []byte) bool {
 	if len(substr) == 0 {
 		return true
@@ -458,7 +461,11 @@ func containsLower(line, substr []byte) bool {
 	for i <= maxIndex {
 		// Find potential first byte match
 		c := line[i]
-		if c != firstByte && c+'a'-'A' != firstByte && c != firstByte+'a'-'A' {
+		// Fast path for ASCII - if c is uppercase letter, convert to lowercase
+		if c >= 'A' && c <= 'Z' {
+			c += 'a' - 'A'
+		}
+		if c != firstByte {
 			i++
 			continue
 		}
@@ -472,9 +479,13 @@ func containsLower(line, substr []byte) bool {
 			c := line[linePos]
 			s := substr[substrPos]
 
-			// Fast ASCII comparison, only check if the character is a letter
-			if c < utf8.RuneSelf && s < utf8.RuneSelf && unicode.IsLetter(rune(c)) && unicode.IsLetter(rune(s)) {
-				if c != s && c+'a'-'A' != s && c != s+'a'-'A' {
+			// Fast path for ASCII
+			if c < utf8.RuneSelf && s < utf8.RuneSelf {
+				// Convert line char to lowercase if needed
+				if c >= 'A' && c <= 'Z' {
+					c += 'a' - 'A'
+				}
+				if c != s {
 					matched = false
 					break
 				}
@@ -485,13 +496,28 @@ func containsLower(line, substr []byte) bool {
 
 			// Slower Unicode path only when needed
 			lr, lineSize := utf8.DecodeRune(line[linePos:])
-			mr, substrSize := utf8.DecodeRune(substr[substrPos:])
+			if lr == utf8.RuneError && lineSize == 1 {
+				// Invalid UTF-8, treat as raw bytes
+				if c >= 'A' && c <= 'Z' {
+					c += 'a' - 'A'
+				}
+				if c != s {
+					matched = false
+					break
+				}
+				linePos++
+				substrPos++
+				continue
+			}
 
-			if lr == utf8.RuneError || mr == utf8.RuneError {
+			mr, substrSize := utf8.DecodeRune(substr[substrPos:])
+			if mr == utf8.RuneError && substrSize == 1 {
+				// Invalid UTF-8 in pattern (shouldn't happen as substr should be valid)
 				matched = false
 				break
 			}
 
+			// Compare line rune converted to lowercase with pattern (which is already lowercase)
 			if unicode.ToLower(lr) != mr {
 				matched = false
 				break